refactor(ax-task): normalize timer module layout
